src/cloud_functions/stage1-power-alert-handler/requirements.txt
Flask==2.3.3
google-cloud-bigquery==3.13.0